requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.8
Django==5.0.6
django-allauth==65.3.1
django-cors-headers==4.4.0
django-haystack==3.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
env==0.1.0
filelock==3.13.1
idna==3.10
packaging==24.1
pillow==10.3.0
platformdirs==4.2.0
PyJWT==2.8.0
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.25.1
Whoosh==2.7.4
